requirements.txt
openai==0.25.0
scikit-learn==1.0.2
tokenizers==0.10.3
transformers==4.11.3
nltk==3.6.5
matplotlib==3.5.3
plotly==5.11.0
simcse==0.4
pytorch-lightning==1.1.5
faiss-cpu==1.7.3
faiss-gpu==1.7.2
torchmetrics==0.9.0
more-itertools==9.0.0
